+++ src/app/join/detail/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import dynamic from 'next/dynamic';
 
-const JoinDetail = dynamic(() => import('@/components/templates/join/detail/JoinDetail'), { ssr: false });
+const JoinDetail = dynamic(() => import('@/components/_pages/join/detail/JoinDetail'), { ssr: false });
 
 export default function JoinDetailPage() {
   return (
