+++ src/app/join/detail/success/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import dynamic from 'next/dynamic';
 
-const JoinDetailSuccess = dynamic(() => import('@/components/templates/join/detail/JoinDetailSuccess'), { ssr: false });
+const JoinDetailSuccess = dynamic(() => import('@/components/_pages/join/detail/JoinDetailSuccess'), { ssr: false });
 
 export default function JoinDetailSuccessPage() {
   return (
