+++ src/app/join/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import dynamic from 'next/dynamic';
 
-const Join = dynamic(() => import('@/components/templates/join/Join'), {
+const Join = dynamic(() => import('@/components/_pages/join/Join'), {
   ssr: false,
 });
