+++ src/app/join/success/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import dynamic from 'next/dynamic';
 
-const JoinSuccess = dynamic(() => import('@/components/templates/join/success/JoinSuccess'), { ssr: false });
+const JoinSuccess = dynamic(() => import('@/components/_pages/join/success/JoinSuccess'), { ssr: false });
 
 export default function JoinSuccessPage() {
   return (
