+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Main } from '@/components/templates';
+import { Main } from '@/components/_pages';
 
 export default function Home() {
   return (
