+++ src/components/_atoms/txt/Txt.tsx
