+++ src/components/_atoms/txt/index.ts
