+++ src/components/features/loginModal/LoginModal.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import styled from '@emotion/styled';
 import CloseRoundedIcon from '@mui/icons-material/CloseRounded';
 
-import { Txt } from '@/components/atoms';
+import { Txt } from '@/components/_atoms';
 import { useModalContext } from '@/contexts/ModalContext';
 import { palette } from '@/styles';
